+++ dao-factory/src/main/java/com/iluwatar/daofactory/UserDAO.java
@@ -1,4 +1,4 @@
-package iluwater.com.daofactory;
+package com.iluwatar.daofactory;
 
 /**
  * Interface that all UserDAOs must support
