+++ dao-factory/src/test/java/com/iluwatar/daofactory/AppTest.java
@@ -1,4 +1,4 @@
-package iluwater.com.daofactory;
+package com.iluwatar.daofactory;
 
 import org.junit.jupiter.api.Test;
